pyproject.toml
[project]
name = "pdf-translator"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohammad",email = "m.khani2810@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfplumber (>=0.11.6,<0.12.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
